Nurogenai - schoolteacher_AI_FE/Nurogenai - schoolteacher_AI/services/api/pyproject.toml
[project]
name = "student-analytics-api"
version = "0.1.0"
description = "FastAPI backend for the Student Analytics Agent Platform"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.15.2",
  "bcrypt<5",
  "email-validator>=2.2.0",
  "fastapi>=0.115.12",
  "google-auth>=2.40.1",
  "httpx>=0.28.1",
  "openai>=1.76.0",
  "passlib[bcrypt]>=1.7.4",
  "pydantic-settings>=2.9.1",
  "pgvector>=0.3.6",
  "psycopg[binary]>=3.2.9",
  "pypdf>=5.4.0",
  "pymupdf>=1.26.0",
  "pytesseract>=0.3.13",
  "requests>=2.32.3",
  "python-jose[cryptography]>=3.5.0",
  "python-multipart>=0.0.20",
  "sqlalchemy>=2.0.40",
  "uvicorn[standard]>=0.34.2"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.5",
  "pytest-cov>=6.1.1"
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]
